+++ src/app/state/hints.rs
@@ -352,9 +352,9 @@ mod tests {
                 RawNode {
                     parent_idx: None,
                     kind: NodeKind::ProcessGroup,
-                    id: "root".into(),
-                    group_id: "root".into(),
-                    name: "root".into(),
+                    id: crate::client::ROOT_GROUP_ID.into(),
+                    group_id: crate::client::ROOT_GROUP_ID.into(),
+                    name: crate::client::ROOT_GROUP_ID.into(),
                     status_summary: NodeStatusSummary::ProcessGroup {
                         running: 0,
                         stopped: 0,
@@ -366,7 +366,7 @@ mod tests {
                     parent_idx: Some(0),
                     kind: NodeKind::ProcessGroup,
                     id: "ingest".into(),
-                    group_id: "root".into(),
+                    group_id: crate::client::ROOT_GROUP_ID.into(),
                     name: "ingest".into(),
                     status_summary: NodeStatusSummary::ProcessGroup {
                         running: 0,
@@ -421,9 +421,9 @@ mod tests {
                 RawNode {
                     parent_idx: None,
                     kind: NodeKind::ProcessGroup,
-                    id: "root".into(),
-                    group_id: "root".into(),
-                    name: "root".into(),
+                    id: crate::client::ROOT_GROUP_ID.into(),
+                    group_id: crate::client::ROOT_GROUP_ID.into(),
+                    name: crate::client::ROOT_GROUP_ID.into(),
                     status_summary: NodeStatusSummary::ProcessGroup {
                         running: 0,
                         stopped: 0,
@@ -435,7 +435,7 @@ mod tests {
                     parent_idx: Some(0),
                     kind: NodeKind::Processor,
                     id: "gen".into(),
-                    group_id: "root".into(),
+                    group_id: crate::client::ROOT_GROUP_ID.into(),
                     name: "Gen".into(),
                     status_summary: NodeStatusSummary::Processor {
                         run_status: "Running".into(),